pyproject.toml
[tool.poetry]
name = "meeting-summariser"
version = "0.1.0"
description = "AI-powered meeting notes summariser with LangGraph agents"
authors = ["Your Name"]
python = "^3.11"

[tool.poetry.dependencies]
python = "^3.11"
google-generativeai = "^0.8.0"
langgraph = "^0.2.0"
langchain-core = "^0.3.0"
langchain-google-genai = "^2.0.0"
chromadb = "^0.5.0"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.31.0"}
python-multipart = "^0.0.12"
openai-whisper = "^20240930"
pyannote-audio = "^3.3.0"
llmlingua = "^0.2.2"
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"
python-dotenv = "^1.0.0"
httpx = "^0.27.0"
redis = "^5.1.0"
tiktoken = "^0.8.0"
numpy = "^1.26.0"

[tool.poetry.dev-dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
ruff = "^0.6.0"
mypy = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.11"
strict = true
